Remove api key from appsettings + reseted changes

+++ frontend/health_app/lib/features/profile/presentation/controllers/profile_onboarding_controller.dart
@@ -123,7 +123,7 @@ class ProfileOnboardingController extends ChangeNotifier {
       years--;
     }
 
-    return years.clamp(0, 130);
+    return years.clamp(0, 130).toInt();
   }
 
   static String? canonicalBloodTypeFromLocalized(String? value) {
